+++ src/Microsoft.Health.CosmosDb/Features/Storage/IDocumentClientInitializer.cs
@@ -3,6 +3,7 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Azure.Documents;
 using Microsoft.Health.CosmosDb.Configs;
@@ -34,7 +35,8 @@ namespace Microsoft.Health.CosmosDb.Features.Storage
         /// </summary>
         /// <param name="documentClient">The <see cref="IDocumentClient"/> instance to use for initialization.</param>
         /// <param name="cosmosDataStoreConfiguration">The data store configuration.</param>
+        /// <param name="collectionInitializers">The collection of collection initializers.</param>
         /// <returns>A task</returns>
-        Task InitializeDataStore(IDocumentClient documentClient, CosmosDataStoreConfiguration cosmosDataStoreConfiguration);
+        Task InitializeDataStore(IDocumentClient documentClient, CosmosDataStoreConfiguration cosmosDataStoreConfiguration, IEnumerable<ICollectionInitializer> collectionInitializers);
     }
 }
