+++ src/Microsoft.Health.CosmosDb/Features/Storage/StoredProcedures/IStoredProcedure.cs
@@ -3,17 +3,17 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-////using System;
-////using Microsoft.Azure.Documents;
+using System;
+using Microsoft.Azure.Documents;
 
-////namespace Microsoft.Health.Fhir.CosmosDb.Features.Storage.StoredProcedures
-////{
-////    internal interface IStoredProcedure
-////    {
-////        string FullName { get; }
+namespace Microsoft.Health.CosmosDb.Features.Storage.StoredProcedures
+{
+    internal interface IStoredProcedure
+    {
+        string FullName { get; }
 
-////        Uri GetUri(Uri collection);
+        Uri GetUri(Uri collection);
 
-////        StoredProcedure AsStoredProcedure();
-////    }
-////}
+        StoredProcedure AsStoredProcedure();
+    }
+}
