+++ src/Microsoft.Health.Fhir.CosmosDb/Features/Storage/StoredProcedures/StoredProcedureInstaller.cs
@@ -3,41 +3,41 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading.Tasks;
-using Microsoft.Azure.Documents;
-using Microsoft.Health.CosmosDb.Features.Storage.Versioning;
+////using System;
+////using System.Collections.Generic;
+////using System.Linq;
+////using System.Net;
+////using System.Threading.Tasks;
+////using Microsoft.Azure.Documents;
+////using Microsoft.Health.CosmosDb.Features.Storage.Versioning;
 
-namespace Microsoft.Health.Fhir.CosmosDb.Features.Storage.StoredProcedures
-{
-    public class StoredProcedureInstaller : ICollectionUpdater
-    {
-        public async Task ExecuteAsync(IDocumentClient client, DocumentCollection collection, Uri relativeCollectionUri)
-        {
-            foreach (IStoredProcedure storedProc in GetStoredProcedures())
-            {
-                try
-                {
-                    await client.ReadStoredProcedureAsync(storedProc.GetUri(relativeCollectionUri));
-                }
-                catch (DocumentClientException e) when (e.StatusCode == HttpStatusCode.NotFound)
-                {
-                    await client.CreateStoredProcedureAsync(relativeCollectionUri, storedProc.AsStoredProcedure());
-                }
-            }
-        }
+////namespace Microsoft.Health.Fhir.CosmosDb.Features.Storage.StoredProcedures
+////{
+////    public class StoredProcedureInstaller : ICollectionUpdater
+////    {
+////        public async Task ExecuteAsync(IDocumentClient client, DocumentCollection collection, Uri relativeCollectionUri)
+////        {
+////            foreach (IStoredProcedure storedProc in GetStoredProcedures())
+////            {
+////                try
+////                {
+////                    await client.ReadStoredProcedureAsync(storedProc.GetUri(relativeCollectionUri));
+////                }
+////                catch (DocumentClientException e) when (e.StatusCode == HttpStatusCode.NotFound)
+////                {
+////                    await client.CreateStoredProcedureAsync(relativeCollectionUri, storedProc.AsStoredProcedure());
+////                }
+////            }
+////        }
 
-        private static IEnumerable<IStoredProcedure> GetStoredProcedures()
-        {
-            var buildInProcs = typeof(IStoredProcedure).Assembly
-                .GetTypes()
-                .Where(x => !x.IsAbstract && typeof(IStoredProcedure).IsAssignableFrom(x))
-                .ToArray();
+////        private static IEnumerable<IStoredProcedure> GetStoredProcedures()
+////        {
+////            var buildInProcs = typeof(IStoredProcedure).Assembly
+////                .GetTypes()
+////                .Where(x => !x.IsAbstract && typeof(IStoredProcedure).IsAssignableFrom(x))
+////                .ToArray();
 
-            return buildInProcs.Select(type => (IStoredProcedure)Activator.CreateInstance(type));
-        }
-    }
-}
+////            return buildInProcs.Select(type => (IStoredProcedure)Activator.CreateInstance(type));
+////        }
+////    }
+////}
